WIP: DX12 RHI

+++ Engine/RHI/Include/RHI/PipelineLayout.h
@@ -5,16 +5,18 @@
 #ifndef EXPLOSION_RHI_PIPELINE_LAYOUT_H
 #define EXPLOSION_RHI_PIPELINE_LAYOUT_H
 
-#include <cstddef>
-
+#include <cstdint>
 #include <Common/Utility.h>
 
 namespace RHI {
     class BindGroupLayout;
 
     struct PipelineLayoutCreateInfo {
-        size_t bindGroupNum;
+        uint32_t bindGroupNum;
         const BindGroupLayout* bindGroupLayouts;
+        // TODO pipeline constant
+        // uint32 pipelineConstantNum;
+        // const PipelineConstantLayout* pipelineConstants;
     };
 
     class PipelineLayout {
